+++ js/router.js
@@ -1,5 +1,5 @@
 import { authService } from "./firebase.js";
-import { viewPost } from "./view-post.js";
+import { viewPost } from "./script-view-post.js";
 
 // export const route = event => {
 //   // 사용자 정보 없으면 a tag의 href 이동 비활성화
@@ -43,7 +43,6 @@ export const handleLocation = async () => {
   }
 
   if (path === "/create-post") {
-    console.log(authService.currentUser);
     //왜인진 모르겠으나 메인을 통해 글쓰기로 와야만 currentUser 데이터를 받아 올 수 있음
     // 따라서 email값이 없으면 메인으로 보내도록 예외처리함
     // 추후 새로고침 해도 currentUser값 받아 올 수 있도록 수정 필요
