+++ pages/view-post.html
@@ -1,6 +1,33 @@
 <div id="page-view-post">
   <div class="container">
     <!-- 댓글 박스-->
+    <article class="post">
+      <div class="post-header">
+        <div class="post-user">
+          <img
+            class="post-profile-img"
+            src="/img/profile-img.png"
+            alt="profile-img"
+          />
+          <div class="post-user-name"></div>
+        </div>
+        <div class="post-create-date"></div>
+      </div>
+      <div class="post-box">
+        <div class="bookmark"><i class="fas fa-mug-hot"></i>0</div>
+        <div class="post-container">
+          <img class="post-img" src="/img/post-img.png" alt="post-img" />
+          <div class="post-content">
+            <div class="title"></div>
+            <div class="description"></div>
+          </div>
+        </div>
+        <div class="post-buttons">
+          <button class="post-modify-btn"></button>
+          <button class="post-delete-btn"></button>
+        </div>
+      </div>
+    </article>
   </div>
 </div>
